requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# PyTorch & CV
torch==2.9.1+cpu
torchvision==0.24.1+cpu
opencv-python==4.12.0.88
pillow==12.0.0

# Data Processing
numpy==2.2.6

# Visualization & Utils
matplotlib==3.10.7
scipy>=1.10.0

# Dependencies for above
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.12.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
packaging==25.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
typing_extensions==4.15.0

# Optional: For production deployment
# gunicorn>=21.2.0
# python-dotenv>=1.0.0
